src/backend/requirements.txt

httpx==0.26.0
python-dotenv==1.0.1
asyncio==3.4.3
